requirements_training.txt
# Training Requirements for Blood Cancer Detection Model
# Install with: pip install -r requirements_training.txt

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Kaggle API
kaggle>=1.5.16

# Additional Training Utils
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Image Augmentation (Optional)
albumentations>=1.3.0

# Model Optimization (Optional)
timm>=0.9.0  # PyTorch Image Models

# GPU Acceleration (Optional - install based on your system)
# For CUDA 11.8: torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: torch torchvision --index-url https://download.pytorch.org/whl/cu121